refactor: Add OfferFeatures and OfferHost.
