+++ src/components/offer-features/offer-features.tsx
@@ -1,12 +1,12 @@
 import { getCapitalizedString } from '../../utils';
 
-type FeaturesProps = {
+type OfferFeaturesProps = {
   type: string;
   bedrooms: number;
   maxAdults: number;
 };
 
-export default function Features({ type, bedrooms, maxAdults }: FeaturesProps): JSX.Element {
+export default function OfferFeatures({ type, bedrooms, maxAdults }: OfferFeaturesProps): JSX.Element {
   return (
     <ul className="offer__features">
       <li className="offer__feature offer__feature--entire">
