--- src/components/premium/premium.tsx
@@ -1,7 +1,0 @@
-export default function Premium(): JSX.Element {
-  return (
-    <div className="place-card__mark">
-      <span>Premium</span>
-    </div>
-  );
-}
