+++ src/utils.ts
@@ -2,4 +2,9 @@ function getCapitalizedString(str: string): string {
   return str[0].toUpperCase() + str.slice(1);
 }
 
-export { getCapitalizedString };
+function getRatingWidth(value: number): string {
+  const widthPercent = (Number.parseInt(value.toFixed(), 10) * 100) / 5;
+  return `${widthPercent}%`;
+}
+
+export { getCapitalizedString, getRatingWidth };
